refactor(cockpit): extract review plan module
